Allow multiple parser implementations
